refactor: shared Repo base + move release domain model to its own package
Go has no classical inheritance, but embedding gets the same practical
result: SkeOperator/Kratix/KratixCli now embed a shared Repo{Source, Name}
instead of each redeclaring an identical Source field and component()
method. Field promotion means m.Source keeps working unchanged.

Moves Component/Artifact/Release/SkeRelease/KratixRelease/Commit/Tag into
their own package, dagger/release, instead of living alongside the Dagger
objects at the module root. This is safe because none of those types are
ever returned from an exported Dagger function (SkeOperator/Kratix/KratixCli
return plain bool/string/error) — Dagger's schema introspection only walks
the module's root package, so types that never cross a Function boundary can
live wherever normal Go package layout says they should.

Re-verified build and is-releasable for all three components after the move.
