+++ dagger/kratix_cli.go
@@ -5,17 +5,14 @@ import (
 	"fmt"
 
 	"dagger/ci/internal/dagger"
+	"dagger/ci/release"
 )
 
 // KratixCli is the kratix-cli component, scoped to a source checkout. No
 // Docker image, no k8s dependency — a plain binary build and a ginkgo suite
 // that fakes out its one external dependency (HELM_BINARY=echo).
 type KratixCli struct {
-	Source *dagger.Directory
-}
-
-func (m *KratixCli) component() Component {
-	return Component{Name: "kratix-cli"}
+	Repo
 }
 
 // Build compiles the kratix CLI binary and returns it as a typed *File — the
@@ -101,6 +98,6 @@ func (m *KratixCli) All(ctx context.Context) (string, error) {
 // define a dedicated KratixCliRelease, and the CLI has no release gate, so
 // Release's default (always true) stands in directly.
 func (m *KratixCli) IsReleasable(ctx context.Context) (bool, error) {
-	release := Release{Component: m.component(), Type: "kratix-cli"}
-	return release.IsReleasable(ctx)
+	rel := release.Release{Component: m.component(), Type: "kratix-cli"}
+	return rel.IsReleasable(ctx)
 }
